+++ CHANGELOG.md
@@ -11,6 +11,9 @@ plus de détails sur http://semver.org/lang/fr/
 
 ## x.β.α (bientôt)
 
+* composant `embed` :
+  * support de l'intégration avec oEmbed
+  * support de l'intégration dans une iframe
 * composant `links` :
   * ajout des effets de rollover
   * support des liens de téléchargements